chore: update filament
