+++ app/Filament/Resources/User/Pages/ListUsers.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Filament\Resources\UserResource\Pages;
+namespace App\Filament\Resources\User\Pages;
 
 use App\Filament\Exports\UserExporter;
-use App\Filament\Resources\UserResource;
+use App\Filament\Resources\User\UserResource;
 use Filament\Actions;
 use Filament\Actions\Exports\Enums\ExportFormat;
 use Filament\Resources\Pages\ListRecords;
