+++ app/Filament/Resources/User/Tables/UsersTable.php
@@ -2,58 +2,22 @@
 
 declare(strict_types=1);
 
-namespace App\Filament\Resources;
+namespace App\Filament\Resources\User\Tables;
 
-use App\Filament\Resources\UserResource\Pages\ListUsers;
-use App\Filament\Resources\UserResource\Pages\ViewUser;
-use App\Models\User;
-use Filament\Forms\Components\DateTimePicker;
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\DeleteAction;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\ViewAction;
 use Filament\Forms\Components\TextInput;
-use Filament\Forms\Form;
-use Filament\Resources\Resource;
-use Filament\Tables\Actions\BulkActionGroup;
-use Filament\Tables\Actions\DeleteAction;
-use Filament\Tables\Actions\DeleteBulkAction;
-use Filament\Tables\Actions\ViewAction;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Enums\FiltersLayout;
 use Filament\Tables\Filters\Filter;
 use Filament\Tables\Table;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Model;
 
-final class UserResource extends Resource
+final readonly class UsersTable
 {
-    protected static ?string $model = User::class;
-
-    protected static ?string $navigationIcon = 'heroicon-o-users';
-
-    #[\Override]
-    public static function canCreate(): bool
-    {
-        return true;
-    }
-
-    #[\Override]
-    public static function canEdit(Model $record): bool
-    {
-        return false;
-    }
-
-    #[\Override]
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                TextInput::make('name')->required(),
-                TextInput::make('email')->required(),
-                DateTimePicker::make('created_at'),
-                DateTimePicker::make('email_verified_at'),
-            ]);
-    }
-
-    #[\Override]
-    public static function table(Table $table): Table
+    public static function configure(Table $table): Table
     {
         return $table
             ->columns([
@@ -63,7 +27,7 @@ final class UserResource extends Resource
             ])
             ->filters([
                 Filter::make('name')
-                    ->form([
+                    ->schema([
                         TextInput::make('name')
                             ->placeholder('Enter name to filter')
                             ->label('Name'),
@@ -74,7 +38,7 @@ final class UserResource extends Resource
                             : $query
                     ),
                 Filter::make('email')
-                    ->form([
+                    ->schema([
                         TextInput::make('email')
                             ->placeholder('Enter email to filter')
                             ->label('Email'),
@@ -86,29 +50,14 @@ final class UserResource extends Resource
                     ),
             ], layout: FiltersLayout::AboveContent)
             ->deferFilters()
-            ->actions([
+            ->recordActions([
                 ViewAction::make()->label(''),
                 DeleteAction::make()->label(''),
             ])
-            ->bulkActions([
+            ->toolbarActions([
                 BulkActionGroup::make([
                     DeleteBulkAction::make(),
                 ]),
             ]);
-    }
-
-    #[\Override]
-    public static function getRelations(): array
-    {
-        return [];
-    }
-
-    #[\Override]
-    public static function getPages(): array
-    {
-        return [
-            'index' => ListUsers::route('/'),
-            'view' => ViewUser::route('/{record}'),
-        ];
     }
 }
