+++ composer.json
@@ -5,7 +5,7 @@
     "type": "project",
     "require": {
         "php": "^8.4",
-        "filament/filament": "^3.3.37",
+        "filament/filament": "^4.0.3",
         "laravel/framework": "^12.25",
         "laravel/pulse": "^1.4.3",
         "olssonm/l5-very-basic-auth": "^7.2",
